[PW-2217] Renaming SalesChannelRepository
