+++ src/Resources/config/services.xml
@@ -10,14 +10,14 @@
 
     <services>
         <!--Adyen Services-->
-        <service id="Adyen\Shopware\Service\Util\SalesChannelUtil" autowire="true">
+        <service id="Adyen\Shopware\Service\Repository\SalesChannelRepository" autowire="true">
             <argument type="service" id="sales_channel_domain.repository"/>
             <argument type="service" id="sales_channel.repository"/>
         </service>
         <service id="Adyen\Shopware\Controller\SalesChannelApiController">
             <argument type="service" id="Adyen\Shopware\Service\OriginKeyService"/>
             <argument type="service" id="Adyen\Shopware\Service\PaymentMethodsService"/>
-            <argument type="service" id="Adyen\Shopware\Service\Util\SalesChannelUtil"/>
+            <argument type="service" id="Adyen\Shopware\Service\Repository\SalesChannelRepository"/>
             <tag name="controller.service_arguments"/>
         </service>
         <service id="Adyen\Shopware\Service\OriginKeyService" autowire="true">
@@ -78,7 +78,7 @@
                       id="Adyen\Shopware\Storefront\Page\Checkout\Confirm\AdyenCheckoutConfirmPageLoader.inner"/>
             <argument type="service" id="Adyen\Shopware\Service\PaymentMethodsService"/>
             <argument type="service" id="Adyen\Shopware\Service\OriginKeyService"/>
-            <argument type="service" id="Adyen\Shopware\Service\Util\SalesChannelUtil"/>
+            <argument type="service" id="Adyen\Shopware\Service\Repository\SalesChannelRepository"/>
             <argument type="service" id="Adyen\Shopware\Service\ConfigurationService"/>
         </service>
         <service id="Adyen\Shopware\Storefront\Page\Checkout\Confirm\AdyenCheckoutConfirmPage"
